+++ src/main/java/vartas/discord/bot/mpi/MPIStatusCode.java
@@ -1,5 +1,5 @@
 package vartas.discord.bot.mpi;
 
 public interface MPIStatusCode {
-    int getCode();
+    short getCode();
 }
